Minor refactorings in EBuildingType's data structures (RelativeBricklayer and RelativeStack).
