+++ jsettlers.common/src/jsettlers/common/buildings/stacks/RelativeStack.java
@@ -12,7 +12,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package jsettlers.common.buildings;
+package jsettlers.common.buildings.stacks;
 
 import jsettlers.common.material.EMaterialType;
 import jsettlers.common.position.RelativePoint;
